additional check for run processor remove raw folder

+++ app/tasks.py
@@ -123,10 +123,11 @@ def copy_processor_local(file_path, copy_back=False):
         for directory in [utl.config_path, utl.raw_path, utl.dict_path]:
             new_path = os.path.join(new_file_path, directory)
             old_path = os.path.join(file_path, directory)
+            if (not copy_back and directory == utl.raw_path and
+                    os.path.exists(new_path)):
+                shutil.rmtree(new_path, ignore_errors=True)
             if not os.path.exists(new_path):
                 os.makedirs(new_path)
-            if not copy_back and directory == utl.raw_path:
-                shutil.rmtree(new_path, ignore_errors=True)
             copy_tree_no_overwrite(old_path, new_path, overwrite=True)
         return new_file_path
     else:
